Epic 6 complete: AI Abstraction & Background Job Infrastructure
Implements E6 (T6.1.1 - T6.3.2): the provider-agnostic AI boundary and the serial
job queue that Writing and Student Problems processing (E7) run on top of, fully
exercised against a fake provider before any real LLM call exists.

T6.1.1 src/ai/AIEvaluationService.ts, src/ai/errors.ts
  The interface (evaluateWriting, processStudentProblemsText) and the failure
  vocabulary. The three error classes carry a `retryable` flag because only the
  code that talked to the provider knows how the call failed; the policy that
  acts on it stays in JobService, where Section 7 puts retry bookkeeping. There
  is deliberately no overall-score field on WritingEvaluation: FR-WRITE-006 makes
  that number the application's, not the model's.

T6.1.2 src/ai/schemas.ts
  Zod schemas for the expected LLM JSON, applied inside the provider so nothing
  downstream re-checks. Each is annotated z.ZodType<T> against the boundary's own
  type, so a schema that stops validating a field the interface promises fails to
  compile — confirmed by deleting one and watching tsc reject it. 20 unit tests.
  Criterion scores are 0-100 but not .int(): integrality matters only for the
  overall score, which is an Int? column, and is WritingScoreCalculator's problem.

T6.1.3 src/test/fakeAIEvaluationService.ts
  A real implementation of the interface, not a stub: it validates its own output
  through the same schemas and raises the same error classes, and refuses to call
  configured malformed output malformed if the schema would have accepted it.
  Scripted responses ("fail twice then succeed") so the retry paths are
  expressible without padding a script to MAX_ATTEMPTS.

T6.2.1 src/data/ProcessingJobRepository.ts
  The claim: UPDATE ... WHERE id = (SELECT ... FOR UPDATE SKIP LOCKED), one
  statement, so there is no window between choosing a row and taking it.
  SKIP LOCKED matters twice over — it is what makes contending claims fan out
  instead of blocking. Proven by removing the mechanism: without it, 8 concurrent
  claims over 8 jobs returned as few as 4 distinct jobs, and 8 claims over 1 job
  returned that job 4 times. Which test caught it varied run to run, so a
  deterministic lock-holding test carries the guarantee (3/3 when SKIP LOCKED
  alone is removed). The claim does not increment attemptCount — Section 8 puts
  that in the failure path, and counting both would double every ordinary failure.

T6.2.2 src/domain/jobs/JobService.ts
  completeJob writes the AI result and the job transition in ONE transaction
  (Section 6 boundary #4); the test provokes a mid-transaction failure and asserts
  both halves stayed unwritten. failJob's retry policy is a parameter
  (maxAttempts + a backoffMs schedule), so no test waits on a clock, and Section
  8's "a schema-validation failure is retryable once" is a two-attempt budget for
  AIValidationError specifically. No failure path names the column holding the
  student's words: retention is a property of the statements, not of discipline.
  JOB_TYPES was MOVED (not copied) from SubmissionService to src/domain/jobs/
  jobTypes.ts, as E5's comment invited — the alternative was JobService and the
  worker importing the submission service, and through it the content loader and
  scoring engine, for two string literals.

T6.2.3 SubmissionRepository.getEvaluationContext()
  Authoritative response text and the frozen contentVersion, resolved from the
  Submission row and nowhere else. Throws rather than returning an empty string
  when the expected field is missing: a blank essay would be *scored*, plausibly,
  with nothing anywhere reporting a problem. 8 integration tests.

T6.3.1 src/background/workerLoop.ts
  Section 8's poll tick with the isRunning guard, which was proven by removing it
  — two ticks fired together then both reached the provider, the one thing the
  single-in-flight-request constraint forbids. tick() is written never to reject:
  its caller is a setInterval callback, and a rejection there is an unhandled
  rejection that ends the process serving the API.

T6.3.2 src/server.ts
  Step 4 of the documented boot order. There is no production AI provider in this
  build — OllamaProvider is E7 (T7.2.3) — so what a production boot does is a
  decision, not an omission: selectAIEvaluationService() is the explicit
  injectable seam (Section 17) and returns null, and on null the API boots, the
  loop does NOT start, and the log says why. Starting it against a provider that
  errors on every call would write failed_needs_review onto real submissions,
  claiming "processing failed, a human should look at this" when the truth is
  "processing was never attempted". Jobs stay pending, which is true. T7.2.3 is a
  one-line change to this function.

  Verified live both ways: unpatched, HTTP 200 with the loop explicitly not
  started; with the seam temporarily returning a fake (run, then reverted), the
  boot log showed the worker started exactly once alongside a single listener, and
  a real submission driven over HTTP through PILOT-2026 moved
  pending/pending -> succeeded/pending (t+3.1s) -> succeeded/succeeded (t+9.9s) on
  the 7s poll. writingOverallScore was still null in the live row, so the E6
  boundary holds outside the suite.

E6 stops short of computing a writing score. completeJob persists the
schema-validated result and moves the status, leaving writingOverallScore null;
WritingScoreCalculator is E7 (T7.1.1), and T7.3.1 is what calls it from the
writing branch. The weights in writing-rubric.json are still PRD Section 23.1
placeholders, and are not E6's to invent. Consequence for E7: the criteria-score,
feedback, and Student Problems derived writes already exist here for both job
types, so T7.3.1 adds the score calculation and T7.4.1's derived write is largely
already in place.

Carried over from E5's review, NOT Epic 6 functionality (staged before this
session began; read, verified, and left as written):

  - contentSchemas.ts: `points` is now z.number().positive().int(). The score
    columns are Int?, so a fractional value used to validate at boot and then make
    Prisma reject the write at submit time — in front of a student, on the one
    action that cannot be retried.

  - rateLimit.ts: a new resetRateLimits(), called by useCleanTestDatabase()
    alongside table truncation. The counters are process-wide by design, so every
    test in a file shared one 60-per-IP budget for POST /api/session/student-verify,
    and a file that grew past 60 failed with 429s that read as an authentication
    bug. The reset clears counts only: it raises no limit, shortens no window, and
    makes nothing skippable, and the limiter's own suites still prove the threshold.
    E6's integration tests draw on a fresh budget per test.

Verification: 344 tests pass across 23 files (from 289 before this epic), the
worker suites exit cleanly with no open handles, `npm run typecheck` is clean for
both halves, and `npm run build` produces dist/ plus the frontend bundle.

+++ src/content/contentSchemas.ts
@@ -66,8 +66,22 @@ export const questionSchema = z
     /** FR-DET-002 — prewritten feedback shown per FR-DET-004, never generated at runtime. */
     explanation: z.string().min(1),
 
-    /** FR-DET-002 — scoring information. */
-    points: z.number().positive(),
+    /**
+     * FR-DET-002 — scoring information: how much this question is worth.
+     *
+     * Whole numbers only, and deliberately so. `Submission.grammarScore`/`vocabularyScore`/
+     * `readingScore` are `Int?` columns (Section 6), and a section's score is the sum of its
+     * questions' points — so a fractional value here would be a content file that validates at boot
+     * and then makes Prisma reject the write at *submission* time, which is the worst moment a
+     * configuration error can surface: in front of a student, on the one action that cannot be
+     * retried.
+     *
+     * `.int()` moves that failure to where the rest of this file's checks already put content
+     * mistakes — boot, before a broken file can reach anyone. If the team later wants partial
+     * credit, the column type and this rule change together in a deliberate migration rather than
+     * the two silently disagreeing.
+     */
+    points: z.number().positive().int(),
   })
   .strict()
   .superRefine((question, ctx) => {
